backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
alembic==1.16.5
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
pydantic-settings==2.10.1
psycopg[binary]==3.2.9
pillow==11.3.0
pytesseract==0.3.13
